+++ Change Log.md
@@ -19,3 +19,11 @@
 - added a new error message for user input
 - added an error message for exernal files that were not imported correctly
 - code displays a positive message if all files are successfully
+
+## Version 5.0
+#### Fifth in-class assignment "RPG Classes"
+- Added a randomly geneating map that can be naviagted by using a s d w
+- added classes for enemies, character and map
+- players can pick up items in the map
+- players can fight monster in the map (this feature is limited at the moment, but will have many features added to it in future)
+- players can choose a name an class for their character